backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pgvector==0.4.1
httpx==0.28.1
python-multipart==0.0.20
agno==1.8.0
openai==1.107.2
chonkie==1.1.1
